pyproject.toml
[project]
name = "python-teams-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.14",
    "asyncio>=3.4.3",
    "botbuilder-core>=4.16.2",
    "cookiecutter==1.7.0",
    "python-dotenv>=1.1.0",
]
